Introduce notebook tests for PRs and run isolated notebook tests only on changed files.
