+++ .github/workflows/ci.yml
@@ -80,10 +80,7 @@ jobs:
           python-version: '3.7'
           architecture: 'x64'
       - name: Install requirements
-        run: |
-          pip install -r requirements.txt
-          python -m pip install -r cirq/contrib/contrib-requirements.txt
-          pip install -r dev_tools/conf/pip-list-dev-tools.txt
+        run: pip install -r requirements.txt -r cirq/contrib/contrib-requirements.txt -r dev_tools/conf/pip-list-dev-tools.txt
       - name: RST check
         run: find . -type f -name "*.rst" | xargs rstcheck --report warning --ignore-directives autoclass,automodule
       - name: Doc check
@@ -206,8 +203,8 @@ jobs:
           pip install -r dev_tools/conf/pip-list-dev-tools.txt
       - name: Pytest check
         run: check/pytest --ignore=cirq/contrib --benchmark-skip
-  notebooks:
-    name: Notebook Tests
+  notebooks-stable:
+    name: Changed Notebooks Isolated Test against Cirq stable
     runs-on: ubuntu-16.04
     steps:
       - uses: actions/checkout@v2
@@ -218,6 +215,29 @@ jobs:
           python-version: '3.7'
           architecture: 'x64'
       - name: Install requirements
-        run: pip install pytest pytest-xdist requests virtualenv filelock
+        run: pip install -r dev_tools/notebooks/requirements-isolated-notebook-tests.txt
       - name: Notebook tests
-        run: check/pytest -n auto -m slow dev_tools/notebook_test.py
+        run: check/pytest -n auto -m slow dev_tools/notebooks/isolated_notebook_test.py
+      - uses: actions/upload-artifact@v2
+        if: failure()
+        with:
+          name: notebook-outputs
+          path: out
+  notebooks-branch:
+    name: Notebook Tests against PR
+    runs-on: ubuntu-16.04
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v1
+        with:
+          python-version: '3.7'
+          architecture: 'x64'
+      - name: Install requirements
+        run: pip install -r requirements.txt -r cirq/contrib/contrib-requirements.txt -r dev_tools/notebooks/requirements-notebook-tests.txt
+      - name: Notebook tests
+        run: check/pytest -n auto -m slow dev_tools/notebooks/notebook_test.py
+      - uses: actions/upload-artifact@v2
+        if: failure()
+        with:
+          name: notebook-outputs
+          path: out
