+++ .gitignore
@@ -46,3 +46,6 @@ docs/api_docs
 
 # pyenv configuration
 .python-version
+
+# notebook test output
+out
