+++ dev_tools/conf/pip-list-dev-tools.txt
@@ -6,7 +6,7 @@ pytest~=5.4.1
 pytest-asyncio~=0.12.0
 pytest-cov~=2.5.0
 pytest-benchmark~=3.2.0
-filelock
+filelock~=3.0.12
 
 # For generating protobufs
