+++ dev_tools/notebooks/requirements-isolated-notebook-tests.txt
@@ -1,0 +1,6 @@
+pytest>=6.0.0
+pytest-xdist~=2.2.0
+filelock~=3.0.12
+virtualenv
+# for shell_tools
+requests
