+++ dev_tools/notebooks/requirements-notebook-tests.txt
@@ -1,0 +1,12 @@
+pytest>=6.0.0
+pytest-xdist~=2.2.0
+filelock~=3.0.12
+
+papermill~=2.3.2
+notebook~=6.2.0
+
+# https://github.com/nteract/papermill/issues/519
+ipykernel==5.3.4
+
+# assumed to be part of colab
+seaborn~=0.11.1
